% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find_Areas.R
\name{findLand}
\alias{findLand}
\title{Find the name of the land on which the occurrence points were found}
\usage{
findLand(occurrences, fillgaps = FALSE)
}
\arguments{
\item{occurrences}{The dataframe output by getData (or if using a custom dataframe, ensure that it has the following columns: decimalLongitude, decimalLatitude, acceptedScientificName)}

\item{fillgaps}{(logical) Attempt to use Photon API to fill in gaps left by mapdata::map.where (TRUE) or only mapdata::map.where results (FALSE, default). While it is powerful, the Photon API does not have a standard location for island names in its returned information, so using it will likely require the returned dataframe to be cleaned by the user.}
}
\value{
A dataframe of the species name, longitude, latitude, country, and island of occurrence
}
\description{
Use various mapping tools to attempt to find the names of land masses where the occurrence points were found.
}
\examples{
\dontrun{occs <- findLand(occurrences, fillgaps = FALSE)}
}
