% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_island_areas.R
\name{get_island_areas}
\alias{get_island_areas}
\title{Access properly encoded island_area object}
\usage{
get_island_areas()
}
\value{
An edited version of the SSARP::island_areas object, which is a
dataframe including the names, areas, and maximum elevations of islands
from across the globe.
}
\description{
In order to address an R CMD check warning about non-ASCII characters in the
island_area object, these characters in the island names had to be converted
to an ASCII format. The non-ASCII accents in the island names are important
for the functionality of the SSARP package, so this function provides the
user with a dataframe including the original, un-converted island names.
}
\examples{
\dontrun{island_df <- get_island_areas()}
}
