% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_presence_absence.R
\name{get_presence_absence}
\alias{get_presence_absence}
\title{Create a presence-absence dataframe for a given occurrence record dataframe}
\usage{
get_presence_absence(occs)
}
\arguments{
\item{occs}{The dataframe output by SSARP::findAreas, or if using a custom
dataframe, ensure that it has the following named columns:
\itemize{
\item "areas" containing the areas associated with the land masses of interest
\item "Species" containing the names of the species living on those islands
\item A 6th column containing locality information. In the SSARP workflow, this
column is called "First" and contains the country name
\item A 7th column containing locality information. In the SSARP workflow, this
column is called "Second" and contains a province or island name
\item An 8th column containing locality information. In the SSARP workflow, this
column is called "Third" and contains the island name if the 7th column does
not contain the island name
}}
}
\value{
A dataframe with a row for each island in the given occurrence
record dataframe and a column for each species. Within each species column,
a 1 represents the presence of that species on the island corresponding to
the given row, and a 0 represents the absence of that species on the island
corresponding to the given row.
}
\description{
Use a dataframe output by SSARP::findAreas to determine which species occur
on specific islands by creating a presence-absence dataframe. A 1 represents
presence and a 0 represents absence.
}
\examples{
# The GBIF key for the Anolis genus is 8782549
#  Obtained with: key <- get_key(query = "Anolis", rank = "genus")
key <- 8782549
# Read in example dataset obtained through:
#  dat <- get_data(key = key, limit = 100)
dat <- read.csv(system.file("extdata",
                            "SSARP_Example_Dat.csv",
                            package = "SSARP"))
land <- find_land(occurrences = dat)
areas <- find_areas(occs = land)
pres_abs <- get_presence_absence(areas)

}
