% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_richness.R
\name{get_richness}
\alias{get_richness}
\title{Create a species richness dataframe for a given occurrence record dataframe}
\usage{
get_richness(occs)
}
\arguments{
\item{occs}{The dataframe output by \code{ssarp::find_areas()}, or if using a
custom dataframe, ensure that it has the following named columns:
\itemize{
\item "areas" containing the areas associated with the land masses of interest
\item "specificEpithet" containing the names of the species living on those
islands
}}
}
\value{
A dataframe with two columns: the first containing island areas and
the second containing the associated species richness (number of unique
species)
}
\description{
Use a dataframe output by \code{ssarp::find_areas()} to determine how many
species occur on each island by creating a species richness dataframe.
}
\details{
The output of this function can be used directly with
\href{https://txm676.github.io/sars/articles/sars-r-package.html}{the sars R package}
to fit additional SAR models that \code{ssarp} does not create itself.
}
\examples{
# The GBIF key for the Anolis genus is 8782549
# Read in example dataset filtered from:
#  dat <- rgbif::occ_search(taxonKey = 8782549, 
#                           hasCoordinate = TRUE,
#                           limit = 10000)
dat <- read.csv(system.file("extdata",
                            "ssarp_Example_Dat.csv",
                            package = "ssarp"))
land <- find_land(occurrences = dat)
areas <- find_areas(occs = land)
richness <- get_richness(occs = areas)

}
