% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gather_Data.R
\name{getData}
\alias{getData}
\title{Get occurrence data from GBIF}
\usage{
getData(key, limit = 100)
}
\arguments{
\item{key}{The taxon key for the taxon of interest (can be found using SSARP::get_key)}

\item{limit}{The maximum number of occurrence records to return. Default: 100. NOTE: there is a hard maximum of 100,000 (see more information about the limit in rgbif::occ_search)}
}
\value{
A dataframe of occurrence records for the taxon of interest
}
\description{
Use rgbif to retrieve occurrence data from GBIF's database for a given taxon. This function will only return occurrences that include GPS coordinates.
}
\examples{
\dontrun{
dat <- getData(key = 2447418, limit = 1000)
}
}
