% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gather_Data.R
\name{getKey}
\alias{getKey}
\title{Retrieve taxon key from GBIF}
\usage{
getKey(query, rank)
}
\arguments{
\item{query}{The search parameter to pass to rgbif's name_suggest() function}

\item{rank}{The taxonimic rank associated with the search parameter}
}
\value{
The taxon key that GBIF needs to search its database for occurrence records
}
\description{
Use rgbif to find the best match for the name given and return the taxon key
}
\examples{
\dontrun{
key <- get_key("Phelsuma", "genus")
key <- get_key("Tamias ruficaudus ruficaudus", "subspecies")
key <- get_key("Naesiotus akamatus", "species")
key <- get_key("Araneae", "order")
}
}
