% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSources.R
\name{getSources}
\alias{getSources}
\title{Gather sources from GBIF data for citation}
\usage{
getSources(occs)
}
\arguments{
\item{occs}{The occurrence record dataframe returned by SSARP::getData or
SSARP::findAreas.}
}
\value{
A dataframe of dataset keys and the number of occurrence records
associated with each key that were gathered with SSARP::getData and/or
filtered with SSARP::findAreas.
}
\description{
When using data obtained via SSARP::getData and filtered with
SSARP::findAreas for a publication, you must keep a record of the datasets
used in your analysis. This function assists in creating the dataframe
necessary to follow GBIF's citation guidelines (see References).
}
\examples{
\dontrun{
key <- getKey(query = "Anolis", rank = "genus")
occs <- getData(key = key, limit = 10000)
source_df <- getSources(occs = occs)
}
}
\references{
\itemize{
\item \href{https://www.gbif.org/citation-guidelines}{GBIF citation guidelines}
\item Data obtained via SSARP::getData and filtered with SSARP::findAreas falls
under \href{https://www.gbif.org/derived-dataset/about}{the derived datasets distinction}
\item \href{https://data-blog.gbif.org/post/derived-datasets/}{More information about creating derived datasets}
}
}
