% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{island_areas}
\alias{island_areas}
\title{Island name data}
\format{
\subsection{'island_areas'}{

A data frame with 26376 rows and 5 variables:
\describe{
\item{\code{OBJECTID_1}}{integer Object ID from ArcGIS}
\item{\code{COUNT}}{integer Number of cells within island, from ArcGIS}
\item{\code{AREA}}{double Area of the island}
\item{\code{MAX}}{integer Maximum elevation of the island}
\item{\code{Name}}{character Name of the island}
}
}
}
\usage{
island_areas
}
\description{
A collection of island names, their areas, and their elevations found using
ArcGIS
}
\keyword{datasets}
