% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.SAR.R
\name{plot.SAR}
\alias{plot.SAR}
\title{Plot a species-area relationship}
\usage{
\method{plot}{SAR}(x, ...)
}
\arguments{
\item{x}{The SAR object that will be plotted}

\item{...}{Functions to pass to plot()}
}
\value{
A plot of your species-area relationship
}
\description{
Function for plotting species-area relationship objects from the
\code{ssarp::create_SAR()} function
}
\examples{
# The GBIF key for the Anolis genus is 8782549
# Read in example dataset filtered from:
#  dat <- rgbif::occ_search(taxonKey = 8782549, 
#                           hasCoordinate = TRUE,
#                           limit = 10000)
dat <- read.csv(system.file("extdata",
                            "ssarp_Example_Dat.csv",
                            package = "ssarp"))
occs <- find_land(occurrences = dat)
areas <- find_areas(occs = occs)
seg <- create_SAR(areas, npsi = 0)
plot(seg)

}
