% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.SAR.R
\name{print.SAR}
\alias{print.SAR}
\title{Print species-area relationship summary}
\usage{
\method{print}{SAR}(x, printlen = NULL, ...)
}
\arguments{
\item{x}{The SAR object of interest}

\item{printlen}{Should always be NULL}

\item{...}{Parameters to pass to print()}
}
\value{
The summary of your species-area relationship
}
\description{
Function for printing the summary of species-area relationship objects from
the \code{ssarp::create_SAR()} function
}
\examples{
# The GBIF key for the Anolis genus is 8782549
# Read in example dataset filtered from:
#  dat <- rgbif::occ_search(taxonKey = 8782549, 
#                           hasCoordinate = TRUE,
#                           limit = 10000)
dat <- read.csv(system.file("extdata",
                            "ssarp_Example_Dat.csv",
                            package = "ssarp"))
occs <- find_land(occurrences = dat)
areas <- find_areas(occs = occs)
seg <- create_SAR(areas, npsi = 0)
print(seg)

}
