% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_create_SAR.R
\name{quick_create_SAR}
\alias{quick_create_SAR}
\title{Quickly create a species-area relationship}
\usage{
quick_create_SAR(
  taxon,
  rank,
  limit = 100,
  geometry = NULL,
  continent = TRUE,
  npsi = 1
)
}
\arguments{
\item{taxon}{The name of the taxon of interest.}

\item{rank}{The taxonomic rank associated with the taxon of interest.}

\item{limit}{The number of GBIF occurrence records to obtain.}

\item{geometry}{(character) A polygon written in Well Known Text (WKT)
format to pass to the SSARP::get_data function. If no polygon is given,
records from across the globe will be returned.}

\item{continent}{(logical) If TRUE, continental areas are removed from the
species-area relationship data. If FALSE, continental areas are not removed
from the species-area relationship data. Default: TRUE}

\item{npsi}{The number of breakpoints to estimate. Default: 1}
}
\value{
A list of 3 including: the summary output, the segmented regression
object, and the aggregated dataframe used to create the plot.
}
\description{
Use the SSARP workflow for creating a species-area relationship for island
species using occurrence data from GBIF without having to go through every
individual step.
}
\examples{
seg <- quick_create_SAR(taxon = "Phelsuma", rank = "genus", limit = 100, npsi = 0)

}
