% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciationDR.R
\name{speciationDR}
\alias{speciationDR}
\title{Get speciation rates using the DR statistic (Jetz et al. 2012)}
\usage{
speciationDR(tree, label_type = "binomial", occurrences)
}
\arguments{
\item{tree}{The dated phylogenetic tree that corresponds with the taxa to be included in a speciation-area relationship}

\item{label_type}{Either "epithet" or "binomial" (default): describes the type of tip label in the provided tree. If "epithet," only the species epithet will be used to match speciation rates to tips in the returned occurrence dataframe. If "binomial," the full species name (including genus) will be used to match speciation rates to tips in the returned occurrence dataframe.}

\item{occurrences}{The occurrence record dataframe output from the SSARP pipeline. If you would like to use a custom dataframe, please make sure that there are columns titled "Genus" and "Species"}
}
\value{
A dataframe that includes speciation rates for each species in the occurrence record dataframe
}
\description{
Use methodology from Sun and Folk (2020) to calculate the DR statistic for each tip of a given tree and output a dataframe for use in SSARP's speciation-area relationship pipeline. This method also removes any species rows without rates (this is most likely to occur when the tree does not have all of the species included in the occurrence record dataframe)
}
\examples{
\dontrun{
key <- getKey(query = "Anolis", rank = "genus")
dat <- getData(key = key, limit = 100)
land <- findLand(occurrences = dat)
areas <- findAreas(occs = land)

# Assuming that the user has a tree file called "anolis.tree"
tree <- ape::read.tree("anolis.tree")

occ_speciation <- speciationDR(tree = tree, label_type = "epithet", occurrences = areas)
}
}
\references{
\itemize{
\item Jetz, W., Thomas, G.H, Joy, J.B., Harmann, K., & Mooers, A.O. (2012). The global diversity of birds in space and time. Nature, 491: 444-448.
\item Sun, M. & Folk, R.A. (2020). Cactusolo/rosid_NCOMMS-19-37964-T: Code and data for rosid_NCOMMS-19-37964 (Version V.1.0). Zenodo. http://doi.org/10.5281/zenodo.3843441
}
}
