% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssarp-package.R
\docType{package}
\name{ssarp-package}
\alias{ssarp}
\alias{ssarp-package}
\title{ssarp: ssarp (Species-/Speciation-Area Relationship Projector)}
\description{
Create Species- and Speciation-Area Relationships for any taxa.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kmartinet/ssarp}
  \item \url{https://kmartinet.github.io/ssarp/}
  \item Report bugs at \url{https://github.com/kmartinet/ssarp/issues}
}

}
\author{
\strong{Maintainer}: Kristen Martinet \email{kmartinet@outlook.com}

}
\keyword{internal}
