% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_SAR.R
\name{create_SAR}
\alias{create_SAR}
\title{Create a species-area relationship (SAR) plot}
\usage{
create_SAR(occurrences, npsi = 1, visualize = FALSE)
}
\arguments{
\item{occurrences}{The dataframe output by \code{ssarp::find_areas()} (or if
using  a custom dataframe, ensure that it has the following columns:
specificEpithet, areas)}

\item{npsi}{The maximum number of breakpoints to estimate for model
selection.  Default: 1}

\item{visualize}{(boolean) Whether the plot should be displayed when the
function is called. Default: FALSE}
}
\value{
A list of 3 including: the summary output, the segmented regression
object, and the aggregated dataframe used to create the plot
}
\description{
Use segmented regression to create a species-area relationship (SAR) plot.
The X axis represents log(island area) and the Y axis represents log(number
of species)
}
\examples{
# The GBIF key for the Anolis genus is 8782549
# Read in example dataset filtered from:
#  dat <- rgbif::occ_search(taxonKey = 8782549, 
#                           hasCoordinate = TRUE,
#                           limit = 10000)
dat <- read.csv(system.file("extdata",
                            "ssarp_Example_Dat.csv",
                            package = "ssarp"))
land <- find_land(occurrences = dat)
areas <- find_areas(occs = land)

seg <- create_SAR(occurrences = areas,
                  npsi = 1,
                  visualize = FALSE)
plot(seg)
summary <- seg[1]
}
