% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_SAR.R
\name{create_SAR}
\alias{create_SAR}
\title{Create a species-area relationship (SAR) plot}
\usage{
create_SAR(occurrences, npsi = 1)
}
\arguments{
\item{occurrences}{The dataframe output by \code{SSARP::find_areas()} (or if using
a custom dataframe, ensure that it has the following columns: Species, areas)}

\item{npsi}{The maximum number of breakpoints to estimate for model
selection.  Default: 1}
}
\value{
A list of 3 including: the summary output, the segmented regression
object, and the aggregated dataframe used to create the plot
}
\description{
Use segmented regression to create a species-area relationship (SAR) plot.
The X axis represents log(island area) and the Y axis represents log(number
of species)
}
\examples{
# The GBIF key for the Anolis genus is 8782549
#  Obtained with: key <- get_key(query = "Anolis", rank = "genus")
key <- 8782549
# Read in example dataset obtained through:
#  dat <- get_data(key = key, limit = 100)
dat <- read.csv(system.file("extdata",
                            "SSARP_Example_Dat.csv",
                            package = "SSARP"))
land <- find_land(occurrences = dat)
areas <- find_areas(occs = land)

seg <- create_SAR(areas)
summary <- seg[1]
}
