% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_SpAR.R
\name{create_SpAR}
\alias{create_SpAR}
\title{Create a speciation-area relationship plot (SpAR)}
\usage{
create_SpAR(occurrences, npsi = 1)
}
\arguments{
\item{occurrences}{The dataframe output by one of SSARP's speciation
methods (\code{SSARP::estimate_BAMM()}, \code{SSARP::estimate_DR()},
\code{SSARP::estimate_MS()}), or if using a custom dataframe, ensure that it
has the following columns: areas, rate}

\item{npsi}{The maximum number of breakpoints to estimate for model
selection.  Default: 1}
}
\value{
A list of 3 including: the summary output, the segmented regression
object, and the aggregated dataframe used to create the plot
}
\description{
Use segmented regression to create a speciation-area relationship plot. The
X axis represents log(island area) and the Y axis represents
log(speciation rate)
}
\examples{
\dontrun{
key <- get_key(query = "Anolis", rank = "genus")
dat <- get_data(key = key, limit = 100)
land <- find_land(occurrences = dat)
areas <- find_areas(occs = land)
occ_speciation <- estimate_BAMM(label_type = "epithet", 
                                 occurrences = areas, 
                                 edata = edata)
seg <- create_SpAR(occurrences = occ_speciation, npsi = 1)
summary <- seg[1]
}
}
