% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_SpAR.R
\name{create_SpAR}
\alias{create_SpAR}
\title{Create a speciation-area relationship plot (SpAR)}
\usage{
create_SpAR(occurrences, npsi = 1, visualize = FALSE)
}
\arguments{
\item{occurrences}{The dataframe output by one of ssarp's speciation
methods (\code{ssarp::estimate_BAMM()}, \code{ssarp::estimate_DR()},
\code{ssarp::estimate_MS()}), or if using a custom dataframe, ensure that it
has the following columns: areas, rate}

\item{npsi}{The maximum number of breakpoints to estimate for model
selection.  Default: 1}

\item{visualize}{(boolean) Whether the plot should be displayed when the
function is called. Default: FALSE}
}
\value{
A list of 4 including: the summary output, the regression
object, the aggregated dataframe used to create the plot, and the AIC scores
used in model selection
}
\description{
Use segmented regression to create a speciation-area relationship plot. The
X axis represents log(island area) and the Y axis represents
log(speciation rate)
}
\details{
If the user would prefer to create their own plot of the
\code{ssarp::create_SpAR()} output, the \code{aggDF} element of the returned list
includes the raw points from the plot created here. They can be accessed
as demonstrated in the Examples section.
}
\examples{
# The GBIF key for the Anolis genus is 8782549
# Read in example dataset filtered from:
#  dat <- rgbif::occ_search(taxonKey = 8782549, 
#                           hasCoordinate = TRUE,
#                           limit = 10000)
dat <- read.csv(system.file("extdata",
                            "ssarp_Example_Dat.csv",
                            package = "ssarp"))
land <- find_land(occurrences = dat)
areas <- find_areas(occs = land)

# Read tree from Patton et al. (2021), trimmed to Caribbean species
tree <- ape::read.tree(system.file("extdata",
                                   "Patton_Anolis_trimmed.tree",
                                   package = "ssarp"))

occ_speciation <- estimate_MS(tree = tree,
                              label_type = "epithet",
                              occurrences = areas)

seg <- create_SpAR(occurrences = occ_speciation, 
                   npsi = 1,
                   visualize = FALSE)
plot(seg)
summary <- seg$summary
model_object <- seg$segObj
points <- seg$aggDF
}
