% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_areas.R
\name{find_areas}
\alias{find_areas}
\title{Find areas of land masses.}
\usage{
find_areas(occs, area_custom = NULL)
}
\arguments{
\item{occs}{The dataframe that is returned by \code{ssarp::find_land()}. If using
a custom occurrence record dataframe, ensure that it has the following
columns: "acceptedScientificName", "genericName", "specificEpithet",
"decimalLongitude", "decimalLatitude", "First", "Second", "Third",
"datasetKey". The "datasetKey" column is important for GBIF records and
identifies the dataset to which the occurrence record belongs. Custom
dataframes without this style of data organization should fill the column
with placeholder values.}

\item{area_custom}{A dataframe including names of land masses and their
associated areas. This dataframe should be provided when the user would like
to bypass using the built-in database of island names and areas. Please
ensure that the custom dataframe includes the land mass's area in column 3
and the name in column 5. (Optional)}
}
\value{
A dataframe of the species name, island name, and island area
}
\description{
Reference a dataset of island names and areas to find the areas of the land
masses relevant to the taxon of interest.
}
\examples{
# The GBIF key for the Anolis genus is 8782549
# Read in example dataset filtered from:
#  dat <- rgbif::occ_search(taxonKey = 8782549, 
#                           hasCoordinate = TRUE,
#                           limit = 10000)
dat <- read.csv(system.file("extdata",
                            "ssarp_Example_Dat.csv",
                            package = "ssarp"))
occs <- find_land(occurrences = dat)
areas <- find_areas(occs = occs)
}
