% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find_Areas.R
\name{findAreas}
\alias{findAreas}
\title{Find areas of land masses.}
\usage{
findAreas(occs)
}
\arguments{
\item{occs}{The dataframe that is returned by SSARP::findLand. If using a custom dataframe, ensure that it has the following columns in order: "Species", "Longitude", "Latitude", "First", "Second", "Third"}
}
\value{
A dataframe of the species name, island name, and island area
}
\description{
Reference a dataset of island names and areas to find the areas of the land masses relevant to the taxon of interest.
}
\examples{
\dontrun{areas <- findAreas(occs)}
}
