% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAreas.R
\name{findAreas}
\alias{findAreas}
\title{Find areas of land masses.}
\usage{
findAreas(occs, area_custom = NULL)
}
\arguments{
\item{occs}{The dataframe that is returned by SSARP::findLand. If using a custom occurrence record dataframe, ensure that it has the following columns in order: "SpeciesName", "Genus", "Species", "Longitude", "Latitude", "First", "Second", "Third"}

\item{area_custom}{A dataframe including names of land masses and their associated areas. This dataframe should be provided when the user would like to bypass using the built-in database of island names and areas. Please ensure that the custom dataframe includes the land mass's area in column 3 and the name in column 5. (Optional)}
}
\value{
A dataframe of the species name, island name, and island area
}
\description{
Reference a dataset of island names and areas to find the areas of the land masses relevant to the taxon of interest.
}
\examples{
\dontrun{areas <- findAreas(occs)}
}
