% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Get occurrence data from GBIF}
\usage{
getData(key, limit = 100, geometry = NULL)
}
\arguments{
\item{key}{The taxon key for the taxon of interest (can be found using
SSARP::getKey)}

\item{limit}{The maximum number of occurrence records to return.
Default: 100. NOTE: there is a hard maximum of 100,000 (see more information
about the limit in rgbif::occ_search)}

\item{geometry}{A counter-clockwise winding order well-known text (WKT)
representation of geometry for use in the rgbif::occ_search call. This
parameter will restrict returned occurrence records to the specified
geometry. More information about valid WKT objects can be found in the
rgbif::occ_search vignette.}
}
\value{
A dataframe of occurrence records for the taxon of interest
}
\description{
Use rgbif to retrieve occurrence data from GBIF's database for a given
taxon. This function will only return occurrences that include
GPS coordinates.
}
\examples{
\dontrun{
dat <- getData(key = 2447418, limit = 1000)

# With geometry that restricts returned Anolis data to a box around Florida
# (Includes some records in Georgia and the Biminis due to the size of 
#  the square)
dat <- getData(key = 8782549, 
              limit = 1000, 
              geometry = 'POLYGON((-87.4 24.3, -79.2 24.3, -79.2 31.1,
               -87.4 31.1, -87.4 24.3))')
}
}
