% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/island_areas.R
\docType{data}
\name{island_areas}
\alias{island_areas}
\title{A list of island names and their areas}
\format{
A dataframe
}
\usage{
data(island_areas)
}
\description{
A list of island names and their areas
}
\examples{
data(island_areas)

}
\keyword{datasets}
