% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{island_areas}
\alias{island_areas}
\title{island_areas}
\format{
A data frame with 26220 rows and 5 variables:
\describe{
  \item{\code{OBJECTID_1}}{integer Object ID from ArcGIS}
  \item{\code{COUNT}}{integer Not sure what this is, from ArcGIS}
  \item{\code{AREA}}{double Area of the island}
  \item{\code{MAX}}{integer Maximum elevation of the island}
  \item{\code{Name}}{character Name of the island} 
}
}
\usage{
island_areas
}
\description{
A dataframe of island names, their areas, and their elevations
}
\keyword{datasets}
