% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.SAR.R
\name{plot.SAR}
\alias{plot.SAR}
\title{Plot a species-area relationship}
\usage{
\method{plot}{SAR}(x, ...)
}
\arguments{
\item{x}{The SAR object that will be plotted}

\item{...}{Functions to pass to plot()}
}
\value{
A plot of your species-area relationship
}
\description{
Function for plotting species-area relationship objects from the
SSARP::create_SAR() function
}
\examples{
# The GBIF key for the Anolis genus is 8782549
#  Obtained with: key <- get_key(query = "Anolis", rank = "genus")
key <- 8782549
# Read in example dataset obtained through:
#  dat <- get_data(key = key, limit = 100)
dat <- read.csv(system.file("extdata",
                            "SSARP_Example_Dat.csv",
                            package = "SSARP"))
occs <- find_land(occurrences = dat)
areas <- find_areas(occs = occs)
seg <- create_SAR(areas, npsi = 0)
plot(seg)

}
