% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.SAR.R
\name{print.SAR}
\alias{print.SAR}
\title{Print species-area relationship summary}
\usage{
\method{print}{SAR}(x, printlen = NULL, ...)
}
\arguments{
\item{x}{The SAR object of interest}

\item{printlen}{Should always be NULL}

\item{...}{Parameters to pass to print()}
}
\value{
The summary of your species-area relationship
}
\description{
Function for printing the summary of species-area relationship objects from the SSARP::SARP() function
}
\examples{
\dontrun{
seg <- SARP(occurrences)
print(seg)
}
}
