% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciationBAMM.R
\name{speciationBAMM}
\alias{speciationBAMM}
\title{Get tip speciation rates from BAMM (Rabosky 2014) analysis}
\usage{
speciationBAMM(label_type = "binomial", occurrences, edata)
}
\arguments{
\item{label_type}{Either "epithet" or "binomial" (default): describes the type of tip label in the tree used for the BAMM analysis. If "epithet," only the species epithet will be used to match speciation rates to tips in the returned occurrence dataframe. If "binomial," the full species name (including genus) will be used to match speciation rates to tips in the returned occurrence dataframe.}

\item{occurrences}{The occurrence record dataframe output from the SSARP pipeline. If you would like to use a custom dataframe, please make sure that there are columns titled "Genus" and "Species"}

\item{edata}{The eventdata object created by using the getEventData() function in BAMMtools}
}
\value{
A dataframe that includes speciation rates for each species in the occurrence record dataframe
}
\description{
Use the BAMMtools package (Rabosky et al. 2014) to extract tip speciation rates from user-supplied BAMM analysis objects.
}
\examples{
\dontrun{
occ_speciation <- speciationBAMM("epithet", occs, edata)
}
}
