% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Speciation_Utils.R
\name{speciationDR}
\alias{speciationDR}
\title{Get speciation rates using the DR statistic (Jetz et al. 2012)}
\usage{
speciationDR(tree, occurrences)
}
\arguments{
\item{tree}{The dated phylogenetic tree that corresponds with the taxa to be included in a speciation-area relationship}

\item{occurrences}{The occurrence record dataframe output from the SSARP pipeline. If you would like to use a custom dataframe, please make sure that there is a column titled "Species"}
}
\value{
A dataframe that includes speciation rates for each species in the occurrence record dataframe
}
\description{
Use the DRstat method from the epm package (Title et al. 2022) and output a dataframe for use in SSARP's speciation-area relationship pipeline. This method also removes any species rows without rates (this is most likely to occur when the tree does not have all of the species included in the occurrence record dataframe)
}
\examples{
\dontrun{
occ_speciation <- speciationDR(tree, occs)
}
}
