% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciationDR.R
\name{speciationDR}
\alias{speciationDR}
\title{Get speciation rates using the DR statistic (Jetz et al. 2012)}
\usage{
speciationDR(tree, label_type = "binomial", occurrences)
}
\arguments{
\item{tree}{The dated phylogenetic tree that corresponds with the taxa to be included in a speciation-area relationship}

\item{label_type}{Either "epithet" or "binomial" (default): describes the type of tip label in the provided tree. If "epithet," only the species epithet will be used to match speciation rates to tips in the returned occurrence dataframe. If "binomial," the full species name (including genus) will be used to match speciation rates to tips in the returned occurrence dataframe.}

\item{occurrences}{The occurrence record dataframe output from the SSARP pipeline. If you would like to use a custom dataframe, please make sure that there are columns titled "Genus" and "Species"}
}
\value{
A dataframe that includes speciation rates for each species in the occurrence record dataframe
}
\description{
Use the DRstat method from the epm package (Title et al. 2022) and output a dataframe for use in SSARP's speciation-area relationship pipeline. This method also removes any species rows without rates (this is most likely to occur when the tree does not have all of the species included in the occurrence record dataframe)
}
\examples{
\dontrun{
occ_speciation <- speciationDR(tree, "epithet", occs)
}
}
