\name{SSARP-package}
\alias{SSARP-package}
\alias{SSARP}
\docType{package}
\title{
  SSARP
}
\description{
  \pkg{SSARP} (Species-/Speciation-Area Relationship Projector) provides a suite of functions to help users create speciation- and species-area relationships for island-dwelling taxa.

  Vignettes for using \pkg{SSARP} to create species- and speciation-area relationships can be found at \url{https://kmartinet.github.io/SSARP/}.
  
  The citation for \pkg{SSARP} is listed below, or can be found using \code{citation("SSARP")}.
}

\author{
  Kristen Martinet
  
  Maintainer: Kristen Martinet <kmartinet@outlook.com>
}

\references{
  Martinet, K.M., Román-Palacios, C., and Harmon, L.J. (2025). SSARP: An R Package for Easily Creating Speices- and Speciation-Area Relationships Using Web Databases. \emph{bioRxiv}, doi:10.1101/2024.12.31.630948
}

\keyword{package}
