% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSARP-package.R
\docType{package}
\name{SSARP-package}
\alias{SSARP}
\alias{SSARP-package}
\title{SSARP: SSARP (Species-/Speciation-Area Relationship Projector)}
\description{
Create Species- and Speciation-Area Relationships for any taxa.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kmartinet/SSARP}
  \item \url{https://kmartinet.github.io/SSARP/}
  \item Report bugs at \url{https://github.com/kmartinet/SSARP/issues}
}

}
\author{
\strong{Maintainer}: Kristen Martinet \email{kmartinet@outlook.com}

}
\keyword{internal}
