% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssarp-package.R
\docType{package}
\name{ssarp-package}
\alias{ssarp}
\alias{ssarp-package}
\title{ssarp: Species-/Speciation-Area Relationship Projector}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Create Species- and Speciation-Area Relationships using occurrence records or presence-absence matrices.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci/ssarp}
  \item \url{https://docs.ropensci.org/ssarp}
  \item Report bugs at \url{https://github.com/ropensci/ssarp/issues}
}

}
\author{
\strong{Maintainer}: Kristen Martinet \email{kmartinet@outlook.com} (\href{https://orcid.org/0000-0002-3905-9507}{ORCID})

Other contributors:
\itemize{
  \item Luke Harmon (\href{https://orcid.org/0000-0002-4985-5750}{ORCID}) [contributor]
  \item Cristian Román-Palacios (\href{https://orcid.org/0000-0003-1696-4886}{ORCID}) [contributor]
  \item Tom Matthews (Dr. Matthews reviewed the package for rOpenSci, see <https://github.com/ropensci/software-review/issues/685>) [reviewer]
  \item Joel Nitta (Dr. Nitta reviewed the package for rOpenSci, see <https://github.com/ropensci/software-review/issues/685>) [reviewer]
}

}
\keyword{internal}
